Delete command: Trying to debug "bye" functions
This commit is trying to fix the bugs in the "Delete" function that was causing an IndexOutOfBoundsException even though there has task inside the TaskList.

This bug causes user unable to delete any task inside the list.

To fix the "Delete" function, I traced the code and trying to discover the part that the wrong index was being used to remove the item from the list since there are only few parts that required the index can cause this issue.

Unfortunately, this bug hasn't been fixed yet.

Let's fix this bug to improve the reliability and stability of this function.